fix uninstall failure when Btrfs is used as the storage driver
`snap remove microk8s` fails as follows if the storage driver is Btrfs.

```
$ sudo snap remove microk8s
error: cannot perform the following tasks:
- Remove data for snap "microk8s" (383) (remove /var/snap/microk8s/common/var/lib/docker/btrfs/subvolumes/<id>: operation not permitted)
```

It's because Btrfs subvolumes can't be deleted by rmdir() syscall.
Although this problem disappear if the linux kernel version is
v4.18 or more, it's better to fix it for older kernel users.