Use newer containerd and runc versions (#350)
* Use snapcraft docker image from beta tag as it builds more often (#342)

* Updating a few December dates to be 2018 instead of 2019 (#343)

* Use the correct ConfigMap for nginx ingress (#328)

By default, the ingress installation seems to use a non-existing ConfigMap and having another one that exists but isn't used is confusing. This switches the controller to use the one that microk8s creates.

* Use kubernetes/ingress-nginx 0.22.0 which fixes cookie stickiness issue (#325)

* Use ClusterFirstWithHostNet for nginx dnsPolicy (#340)

The nginx ingress pod is created with `hostNet: true`. This means that
it does not, by default, have access to kuebdns. We need that so that
the ingress controller can map things like ExternalName services to
internal services.

This fixes the problem by setting the dnsPolicy to
ClusterFirstWithHostNet.

Fixes #339

* Use newer containerd and runc versions