Updated to use CoreDNS (#491)
* Updated to use CoreDNS

The 'dns' addon has been changed to use CoreDns by default instead
of kube-dns. This brings it in line with the CoreDNS being the
recommend default with Kubernetes.

Relates to: #271

After feedback from the first commit the following have been changed:
* CoreDNS k8s artefcats have been added to there own yaml and not
the existing dns.yaml
* The existing dns yaml stays the same
* When you enable the dns addon it will apply the new coredns.yaml
* When you disable the dns addon it will look for running pods and
decide whether to remove dns.yaml or coredns.yaml

When doing an upgrade where the dns addon has been previously enabled
it will show as disabled on `microk8s.status` and kube-dns will be
running. To upgrade the user will be required to do the following:
1. `microk8s.disable dns`
2. `microk8s.enable dns`

Relates to: #271