Update sed commands to use ';' delimiter
The previous implementation which used '/' as the delimiter meant that
replacing lines in files that contain '/' (such as file paths) would
fail because the sed expression would be terminated earlier than
expected. Change it to use something that's less likely to cause trouble
when configuring config options.

The alternative is escaping in all of the users, which I opted to avoid
because there's already examples in the tree where plugins attempt to
use this function with paths, and also an upcoming commit will add an
additional example of this.