Enable token authentication on kubelet
After enabling the Prometheus module, I checked the status of the targets, and the status of the kubelet endpoint appears to be Down. (401 Unauthorized)

The token authentication appears to have been blocked with the --client-ca-file set up following the #323 modification.
Adding the --authentication-token-webbook option to /var/snap/microk8s/current/args/kubelet changes the status to Up.
I think this is also related to Pod's CPU Usage not being displayed on the dashboard.

So I hope you will review the PR and add it as a basic option if there is no problem.