Fix path to openssl
Running microk8s on certain distributions such as Centos 7.4, resulted
in failure to execute openssl. This happend because the path to
openssl was not being correctly set within the snap. As a result
either no openssl was found or the worng one was selected leading
to linker relocation errors.

This commit fixes the issue by specifying the full path to openssl.