Migrate Cilium to install/uninstall via new CLI
The Cilium community has developed a new CLI for installing, operating
and debugging Kubernetes clusters with Cilium installed[0]. We can reuse
that CLI here for a much simpler, less error-prone upgrade process.

For the most part, the new CLI almost removes the need for these
microk8s-specific scripts. However, the CLI cannot know how to uninstall
existing CNIs in a microk8s-specific environment, so we still need some
logic here to clean that up in order for Cilium to successfully deploy
and avoid conflicts with existing components.

A few notes about implementation choices:
* Remove previous CNIs and CNI config before initiating Cilium install
  to ensure that Cilium manages pods after the cilium install command
  has been run.
* disable.cilium.sh is handling disabling the old version during
  microk8s upgrade.
* Use 'rm -rf -- $directories" to reduce likelihood of removing bad
  directories, rather than 'rm -rf $directories'

[0] https://github.com/cilium/cilium-cli/

Signed-off-by: Joe Stringer <joe@cilium.io>