Always regenerate csr.conf
Since commit 82683e41282aee0153f6818591e92037db7fd3b0, the tool
refresh_certs.py needs to be told which certificate to refresh.
However when asking to refresh the front proxy client or the server
certificate, the csr.conf is read as is without csr.conf.template being
taken into consideration contrary to what is done when all the
certificates are renewed.
This fixes this inconsistency by always regenerating the file csr.conf.