Only verify partial CA chain
If an intermediate CA is used, we cannot verify the whole chain since
the CA needs to be self-signed for `openssl verify` to work.

We only want to verify that the server certificate is actually issued by
the parent CA (either a root CA or intermediate), thus verifying the
partial chain is sufficient.

See also: https://stackoverflow.com/a/60831762