Increase token size to 32 bytes
This fix bumps the token size from 16 to 32 bytes in order
to match the default minimum entropy as recommended by Python
docs.

Note, that there is also a comparison to a constant TOKEN_ΜΙΝ_LEN
where one module assumed that length to be characters and another
module (utils.py) assumes that length to be bytes. So this fix
uses bytes as the meaning from now on.

Fixes #4481

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>