Use the "run-all-tests" label to set "UNDER_TIME_PRESSURE"
There's an env variable that controls whether to run all tests or not.

In some cases, we may want to run all the tests to ensure that a
certain PR doesn't introduce regressions before actually merging it.

As such, we're introducing a new label called "run-all-tests". If
set, it will set the "UNDER_TIME_PRESSURE" env variable to "false".

While at it, we're making the variable case insensitive, making it
easier to use with GH action expressions such as "contains", used
to check if the "run-all-test" label is set.