Always copy over the default hooks
At the moment, the default hooks are not updated when refreshing
the snap.

There's an explicit check so this was most probably by design so that
custom hooks would not get overriden. However, it means that bug
fixes are not applied automatically and the users need to copy the
files manually, which is undocumented and not very user friendly.

This change will always refresh the default hooks, replacing
the existing files. Custom hooks are expected to use separate files
so that they won't get overriden during upgrades.